+++ app/libraries/Jentil/Setups/Styles/Core.php
@@ -11,7 +11,7 @@ final class Core extends AbstractStyle
     {
         parent::__construct($jentil);
 
-        $this->id = $this->app->theme->stylesheet;
+        $this->id = $this->app->meta['slug'];
     }
 
     public function run()
@@ -24,17 +24,17 @@ final class Core extends AbstractStyle
      */
     public function enqueue()
     {
-        if (\is_rtl()) {
-            $file = '/dist/styles/core-rtl.min.css';
-        } else {
-            $file = '/dist/styles/core.min.css';
-        }
+        $file_system = $this->app->utilities->fileSystem;
+
+        $file = \is_rtl() ?
+            '/dist/styles/core-rtl.min.css' :
+            '/dist/styles/core.min.css';
 
         \wp_enqueue_style(
             $this->id,
-            $this->app->utilities->fileSystem->dir('url', $file),
+            $file_system->dir('url', $file),
             [$this->app->setups['Styles\Normalize']->id],
-            \filemtime($this->app->utilities->fileSystem->dir('path', $file))
+            \filemtime($file_system->dir('path', $file))
         );
     }
 }