+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 0.9.1 - 2018-10-04
+
+### Changed
+- Rename `LICENSE.md` to `LICENSE`
+
+### Fixed
+- Replace `Jentil::$theme` with `Jentil::$meta`, since `wp_get_theme()` does not work if Jentil installed as package.
+
 ## 0.9.0 - 2018-09-27
 
 ### Added